UF-15316: Add integration method to fetch casePdf
