+++ src/main/java/se/sundsvall/oepintegrator/integration/opene/rest/OpeneRestIntegration.java
@@ -2,15 +2,24 @@ package se.sundsvall.oepintegrator.integration.opene.rest;
 
 import static java.time.format.DateTimeFormatter.ISO_LOCAL_DATE;
 import static java.util.Optional.ofNullable;
+import static org.springframework.http.HttpHeaders.CONTENT_DISPOSITION;
+import static org.springframework.http.HttpHeaders.CONTENT_LENGTH;
+import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
+import static org.springframework.http.HttpHeaders.LAST_MODIFIED;
+import static org.zalando.problem.Status.INTERNAL_SERVER_ERROR;
 import static se.sundsvall.oepintegrator.integration.opene.soap.model.message.WebmessageMapper.toWebmessageAttachmentData;
 import static se.sundsvall.oepintegrator.integration.opene.soap.model.message.WebmessageMapper.toWebmessages;
 import static se.sundsvall.oepintegrator.service.mapper.CaseMapper.toCaseEnvelopeList;
 import static se.sundsvall.oepintegrator.utility.Constants.OPEN_E_DATE_TIME_FORMAT;
 
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Set;
 import org.springframework.stereotype.Component;
+import org.zalando.problem.Problem;
 import se.sundsvall.oepintegrator.api.model.cases.CaseEnvelope;
 import se.sundsvall.oepintegrator.api.model.webmessage.Webmessage;
 import se.sundsvall.oepintegrator.api.model.webmessage.WebmessageAttachmentData;
@@ -19,7 +28,7 @@ import se.sundsvall.oepintegrator.utility.enums.InstanceType;
 
 /**
  * Integration class for Open-E REST API. This will be used to call the Open-E REST API like this:
- * 
+ *
  * <pre>
  * {@code
  * public void doStuff(final String municipalityId, final String instanceId) {
@@ -60,7 +69,27 @@ public class OpeneRestIntegration {
 		return toCaseEnvelopeList(client.getCaseListByFamilyId(familyId, status, formatLocalDate(fromDate), formatLocalDate(toDate)).orElse(new byte[0]));
 	}
 
-	private String formatLocalDate(LocalDate localDate) {
+	public void getCasePdfByFlowInstanceId(final String municipalityId, final InstanceType instanceType, final String flowInstanceId, final HttpServletResponse response) {
+		final var client = clientFactory.getRestClient(municipalityId, instanceType);
+		try (final var feignResponse = client.getCasePdfByFlowInstanceId(flowInstanceId);
+			final var inputStream = feignResponse.body().asInputStream();
+			final var outputStream = response.getOutputStream()) {
+
+			final var headerSet = Set.of(LAST_MODIFIED, CONTENT_TYPE, CONTENT_DISPOSITION, CONTENT_LENGTH);
+
+			headerSet.forEach(header -> feignResponse.headers()
+				.getOrDefault(header, List.of())
+				.stream()
+				.findFirst().ifPresent(value -> response.setHeader(header, value)));
+
+			inputStream.transferTo(outputStream);
+			outputStream.flush();
+		} catch (final IOException e) {
+			throw Problem.valueOf(INTERNAL_SERVER_ERROR, "Unable to get case pdf");
+		}
+	}
+
+	private String formatLocalDate(final LocalDate localDate) {
 		return ofNullable(localDate).map(date -> date.format(ISO_LOCAL_DATE)).orElse(null);
 	}
 }
